Add support for push, pop static commands
This requires passing the filename to MemoryAccessCommandTranslator as
the instructions for implementing the vm translator say this is the
expected format.

It would be essential to differentiate between static
variables from different files if this program supported running on
multiple files simultaneously. Currently it does not but that's one of
the requirements for this project.